hhhha

+++ 0x0B-malloc_free/2-str_concat.c
@@ -20,8 +20,8 @@ if (s2 == NULL)
 s2 = "";
 }
 
-size_t len1 = strlen(s1);
-size_t len2 = strlen(s2);
+int len1 = strlen(s1);
+int len2 = strlen(s2);
 
 char *result = malloc(len1 + len2 + 1);
